requirements.txt
Flask==3.0.3
tensorflow==2.17.0
numpy==1.23.5
pillow==11.0.0
opencv-python==4.10.0.84
